separate out store apis, have download
