+++ content_store/lib/content_store.dart
@@ -14,6 +14,7 @@ export 'src/models/broken_image.dart';
 export 'src/models/share_files.dart' show ShareManager;
 export 'src/models/shimmer.dart';
 export 'src/models/store_updater.dart';
+export 'src/models/store_updter_ext_store.dart';
 export 'src/providers/db_store.dart';
 export 'src/storage_service/extensions/ext_file.dart' show ExtFileAnalysis;
 export 'src/storage_service/widgets/get_device_directories.dart'
