+++ content_store/lib/src/providers/store_updater.dart
@@ -6,6 +6,7 @@ import 'package:flutter_riverpod/flutter_riverpod.dart';
 
 import '../models/store_updater.dart';
 import 'db_store.dart';
+import 'downloader_provider.dart';
 
 class StoreUpdaterNotifier extends AsyncNotifier<StoreUpdater> {
   @override
@@ -13,12 +14,20 @@ class StoreUpdaterNotifier extends AsyncNotifier<StoreUpdater> {
     final store = await ref.watch(storeProvider.future);
     final directories = await ref.watch(deviceDirectoriesProvider.future);
     final server = ref.watch(serverProvider);
+    final downloader = ref.watch(downloaderProvider);
     return StoreUpdater(
       store: store,
       directories: directories,
       server: server.identity,
+      downloader: downloader,
     );
   }
+
+/*   sync() {
+    state.whenOrNull(data: (storeUpdater) {
+      state = state
+    });
+  } */
 
   void onRefresh() {}
 }
